Move `CheckBasis` and `CheckCtx` to a new `run_cfg` module
These are used more places than just test traits, so this new module
should be a better home. `run_cfg` will also be expanded in the near
future.
